+++ build_runner_core/lib/src/generate/single_step_reader_writer.dart
@@ -451,6 +451,8 @@ class SingleStepReaderWriter extends AssetReader
 
     var node = _runningBuild.assetGraph.get(id);
     if (node == null) {
+      // Add to the graph for input tracking.
+      _runningBuild.assetGraph.add(AssetNode.missingSource(id));
       return PhasedValue.fixed('');
     }